#1730 put user icon on same row as heading in home page
