+++ src/frontend/src/layouts/PageTitle/NavUserMenu.tsx
@@ -3,7 +3,7 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Link as RouterLink, useHistory } from 'react-router-dom';
 import { GoogleLogout } from 'react-google-login';
 import IconButton from '@mui/material/IconButton';
@@ -18,15 +18,30 @@ import ListItemIcon from '@mui/material/ListItemIcon';
 import SettingsIcon from '@mui/icons-material/Settings';
 import LogoutIcon from '@mui/icons-material/Logout';
 import { canAccessAdminTools } from '../../utils/users';
-import { useTheme } from '@mui/system';
+import { Box, useTheme } from '@mui/system';
+import { Typography } from '@mui/material';
+import { fullNamePipe } from '../../utils/pipes';
+import { useCurrentUser } from '../../hooks/users.hooks';
 
-const NavUserMenu: React.FC = () => {
+interface NavUserMenuProps {
+  widthLimit?: number;
+}
+
+const NavUserMenu: React.FC<NavUserMenuProps> = (widthLimit?) => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const history = useHistory();
   const auth = useAuth();
+  const user = useCurrentUser();
 
   const handleMenu = (event: React.MouseEvent<HTMLElement>) => setAnchorEl(event.currentTarget);
   const handleClose = () => setAnchorEl(null);
+
+  const [width, setWidth] = useState(window.innerWidth);
+  useEffect(() => {
+    window.addEventListener('resize', () => {
+      setWidth(window.innerWidth);
+    });
+  });
 
   const googleAuthClientId = import.meta.env.VITE_REACT_APP_GOOGLE_AUTH_CLIENT_ID;
 
@@ -73,7 +88,18 @@ const NavUserMenu: React.FC = () => {
   const theme = useTheme();
 
   return (
-    <>
+    <Box display="flex" flexDirection={'row'} marginLeft={'auto'} marginBottom={1} width={'auto'}>
+      <Typography
+        variant="body1"
+        marginY={0.5}
+        marginLeft={'auto'}
+        marginRight={1}
+        sx={{
+          color: theme.palette.text.primary
+        }}
+      >
+        {width < (widthLimit?.widthLimit || 600) ? '' : fullNamePipe(user)}
+      </Typography>
       <IconButton
         size="large"
         aria-label="account of current user"
@@ -122,7 +148,7 @@ const NavUserMenu: React.FC = () => {
         {canAccessAdminTools(auth.user) && <AdminTools />}
         {import.meta.env.MODE === 'development' ? <DevLogout /> : <ProdLogout />}
       </Menu>
-    </>
+    </Box>
   );
 };
